+++ src/interfaces/Complexity.ts
@@ -1,0 +1,3 @@
+import { ComplexityEstimator } from "graphql-query-complexity";
+
+export type Complexity = ComplexityEstimator | number;